chore: improve Dockerfile and justfile
This updates how the Dockerfile uses uv to set up the Python virtual
environment. This moves the virtual environment to /venv.

This gets rid of the utility service in favor of mounting volumes
directly in the docker compose lines.

This adds documentation to justfile recipes as well as add list-outdated
and update-lock for managing dependencies.