chore: add build-test-push workflow (#1)
* chore: add build-test-push workflow

This adds a build-test-push workflow that builds the Docker image,
validates it, and then pushes it to GAR.

The "pushes it to GAR" part is commented out until we have settings in
place.

* chore: fix just --version call

* chore: adjust order of docker compose build options

Docker said that --progress is now a general optionand not an option of
the build subcommand. This moves it up.

* chore: fix linting recipes to not run linked services