feat: add github response caching (#30)
This reworks the internals to use httpx instead of response (and
pytest-httpx instead of responses) and adds support for the following:

1. retry on HTTP request errors like HTTP 429
2. LRU caching of HTTP responses with the key being (user, repo,
   from_sha)

That'll use the contents from the cache rather than hit github
repeatedly and contents will naturally age out. maxsize is the default
which is 128 which is fine for now, but we might want to adjust that
some day.